conv op implementation.
will add more test cases in a separate PR - need to convert more tests from caffe2 generated ops tests (they pass on my dev box locally)
and add test cases for auto_pad.

+++ lotus/core/providers/cpu/math/gemm.h
@@ -87,9 +87,9 @@ class Gemm final : public OpKernel {
     math::Gemm<T_X, CPUMathUtil>(
         transA_,
         transB_,
-        static_cast<int>(M),
-        static_cast<int>(N),
-        static_cast<int>(K),
+        M,
+        N,
+        K,
         alpha_,
         X->template data<T_X>(),
         W->template data<T_W>(),
